# backend/public/openapi.json
{
    "openapi": "3.0.0",
    "info": {
        "version": "1.0.0",
        "title": "New Life Nursery",
        "description": "New Life Nursery is a wholesale plant distributor, located in Southern New Jersey. We strive to provide beautiful, healthy plants at competitive prices.",
        "termsOfService": "https://newlifenurseryinc.com/terms-and-conditions",
        "license": {
            "name": "MIT"
        },
        "contact": {
            "email": "newlifenurseryinc@gmail.com"
        }
    },
    "servers": [
        {
            "url": "http://newlifenurseryinc.com/v1"
        }
    ],
    "paths": {
...
        "/register": {
            "put": {
                "summary": "Create a new account",
                "operationId": "registerUser",
                "tags": [
                    "auth"
                ],
                "parameters": [
                    {
                        "name": "first_name",
                        "in": "header",
                        "description": "User's first name",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "John"
                        }
                    },
                    {
                        "name": "last_name",
                        "in": "header",
                        "description": "User's last name",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "Smith"
                        }
                    },
                    {
                        "name": "pronouns",
                        "in": "header",
                        "description": "Pronouns the user identifies as",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "they/them"
                        }
                    },
                    {
                        "name": "business",
                        "in": "header",
                        "description": "Name of business associated with the user",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "Main Street Plants"
                        }
                    },
                    {
                        "$ref": "#/components/parameters/Email"
                    },
                    {
                        "$ref": "#/components/parameters/Password"
                    },
                    {
                        "name": "receives_delivery_updates",
                        "in": "header",
                        "description": "Indicates if the user would like to receive email updates",
                        "required": true,
                        "schema": {
                            "type": "boolean",
                            "example": true
                        }
                    },
                    {
                        "name": "existing_customer",
                        "in": "header",
                        "description": "Indicates if the user has previously ordered from New Life Nursery without an account",
                        "required": true,
                        "schema": {
                            "type": "boolean",
                            "example": true
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Details about the current user session",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Session"
                                }
                            }
                        }
                    },
                    "default": {
                        "$ref": "#/components/responses/UnexpectedError"
                    }
                }
            }
        },
        "/login": {
            "put": {
                "summary": "Log in to existing account",
                "operationId": "loginUser",
                "tags": [
                    "auth"
                ],
                "parameters": [
                    {
                        "$ref": "#/components/parameters/Email"
                    },
                    {
                        "$ref": "#/components/parameters/Password"
                    },
                    {
                        "name": "verificationCode",
                        "in": "header",
                        "description": "Code sent to email to verify account. Only needs to be done once",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Details about the current user session",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "user": {
                                            "$ref": "#/components/schemas/Customer"
                                        },
                                        "session": {
                                            "$ref": "#/components/schemas/Session"
                                        },
                                        "emailVerified": {
                                            "type": "boolean"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "default": {
                        "$ref": "#/components/responses/UnexpectedError"
                    }
                }
            }
        },
        "/reset-password": {
            "get": {
                "summary": "Sends a password reset email",
                "operationId": "sendResetPassword",
                "tags": [
                    "auth"
                ],
                "parameters": [
                    {
                        "$ref": "#/components/parameters/Email"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "default": {
                        "$ref": "#/components/responses/UnexpectedError"
                    }
                }
            },
            "put": {
                "summary": "Sets a new password for a user",
                "operationId": "resetPassword",
                "tags": [
                    "auth"
                ],
                "parameters": [
                    {
                        "name": "code",
                        "in": "header",
                        "description": "Password reset code sent to the user's email",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "$ref": "#/components/parameters/Password"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "default": {
                        "$ref": "#/components/responses/UnexpectedError"
                    }
                }
            }
        },
        "/customers": {
            "get": {
                "summary": "Details about all customers. Must be an admin",
                "operationId": "getCustomers",
                "tags": [
                    "admin"
                ],
                "parameters": [
                    {
                        "$ref": "#/components/parameters/Session"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Details about all customers",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Customer"
                                    }
                                }
                            }
                        }
                    },
                    "default": {
                        "$ref": "#/components/responses/UnexpectedError"
                    }
                }
            }
        },
...
        "/orders": {
            "get": {
                "summary": "Array of order data with specified status",
                "operationId": "getOrders",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/Session"
                    },
                    {
                        "name": "orderStatus",
                        "in": "header",
                        "description": "order status",
                        "required": true,
                        "schema": {
                            "$ref": "#/components/schemas/OrderStatus"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Array of order data",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Order"
                                    }
                                }
                            }
                        }
                    },
                    "default": {
                        "$ref": "#/components/responses/UnexpectedError"
                    }
                }
            }
        },
...
        "/availability": {
            "put": {
                "summary": "Upload a new availability file. Must be an admin",
                "operationId": "uploadGallery",
                "tags": [
                    "admin"
                ],
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "data": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "default": {
                        "$ref": "#/components/responses/UnexpectedError"
                    }
                }
            }
        },
...
    },
    "components": {
        "parameters": {
...
            "Session": {
                "name": "session",
                "in": "cookie",
                "description": "current user session",
                "required": true,
                "schema": {
                    "$ref": "#/components/schemas/Session"
                }
            },
            "Email": {
                "name": "email",
                "in": "header",
                "description": "User's email address",
                "required": true,
                "schema": {
                    "$ref": "#/components/schemas/Email"
                }
            },
            "Password": {
                "name": "password",
                "in": "header",
                "description": "User's new password",
                "required": true,
                "schema": {
                    "type": "string",
                    "example": "hunter123"
                }
            }
        },
        "schemas": {
            "Session": {
                "type": "object",
                "required": [
                    "tag",
                    "token"
                ],
                "properties": {
                    "tag": {
                        "type": "string"
                    },
                    "token": {
                        "type": "string"
                    }
                }
            },
...
            "Customer": {
                "type": "object",
                "required": [
                    "id"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "tag": {
                        "type": "string"
                    },
                    "first_name": {
                        "type": "string"
                    },
                    "last_name": {
                        "type": "string"
                    },
                    "pronouns": {
                        "type": "string"
                    },
                    "business": {
                        "type": "string"
                    },
                    "theme": {
                        "type": "string"
                    },
                    "existing_customer": {
                        "type": "boolean"
                    },
                    "roles": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Role"
                        }
                    },
                    "emails": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Email"
                        }
                    },
                    "phones": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Phone"
                        }
                    },
                    "cart_id": {
                        "type": "integer"
                    },
                    "cart_size": {
                        "type": "integer"
                    }
                }
            },
            "Order": {
                "type": "object",
                "required": [
                    "id"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "status": {
                        "$ref": "#/components/schemas/OrderStatus"
                    },
                    "special_instructions": {
                        "type": "string"
                    },
                    "is_delivery": {
                        "type": "boolean"
                    },
                    "desired_delivery_date": {
                        "type": "number"
                    },
                    "items": {
                        "type": "array"
                    },
                    "delivery_address": {
                        "$ref": "#/components/schemas/Address"
                    },
                    "customer": {
                        "$ref": "#/components/schemas/Customer"
                    }
                }
            },
            "OrderStatus": {
                "type": "integer",
                "enum": [
                    -4,
                    -3,
                    -2,
                    -1,
                    0,
                    1,
                    2,
                    3,
                    4,
                    5
                ]
            },
            "Role": {
                "type": "object",
                "properties": {
                    "title": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    }
                }
            },
            "Email": {
                "type": "object",
                "properties": {
                    "email_address": {
                        "type": "string",
                        "example": "somerandomemail@gmail.com"
                    },
                    "receives_delivery_updates": {
                        "type": "boolean",
                        "example": true
                    }
                }
            },
            "Phone": {
                "type": "object",
                "required": [
                    "unformatted_number"
                ],
                "properties": {
                    "unformatted_number": {
                        "type": "string"
                    },
                    "country_code": {
                        "type": "string"
                    },
                    "extension": {
                        "type": "string"
                    },
                    "is_mobile": {
                        "type": "boolean"
                    },
                    "receives_delivery_updates": {
                        "type": "boolean"
                    }
                }
            },
            "Address": {
                "type": "object",
                "required": [
                    "administrative_area",
                    "locality",
                    "postal_code",
                    "throughfare"
                ],
                "properties": {
                    "tag": {
                        "type": "string"
                    },
                    "name": {
                        "type": "string"
                    },
                    "country": {
                        "type": "string",
                        "minLength": 2,
                        "maxLength": 2,
                        "example": "US"
                    },
                    "administrative_area": {
                        "type": "string"
                    },
                    "sub_administrative_area": {
                        "type": "string"
                    },
                    "locality": {
                        "type": "string"
                    },
                    "postal_code": {
                        "type": "string"
                    },
                    "throughfare": {
                        "type": "string"
                    },
                    "premise": {
                        "type": "string"
                    },
                    "delivery_instructions": {
                        "type": "string"
                    }
                }
            },
...
            "Error": {
                "type": "object",
                "required": [
                    "code",
                    "message"
                ],
                "properties": {
                    "code": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "msg": {
                        "type": "string"
                    }
                }
            }
        },
        "responses": {
            "UnexpectedError": {
                "description": "unexpected error",
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/Error"
                        }
                    }
                }
            }
        }
    }
}